+++ tf/enterpriseproj.tf
@@ -1,0 +1,16 @@
+data "huaweicloud_enterprise_project" "ep_dev" {
+  name = "DEV"
+}
+
+data "huaweicloud_enterprise_project" "ep_stg" {
+  name = "STG"
+}
+
+data "huaweicloud_enterprise_project" "ep_prod" {
+  name = "PROD"
+}
+
+data "huaweicloud_enterprise_project" "ep_net" {
+  name = "NET"
+}
+